+++ Makefile
@@ -24,4 +24,3 @@ run:
 
 %:
 	@:
-
