+++ classes_to_be_readded/player.cpp
@@ -30,24 +30,24 @@ void Player::initVariables() {
     */
 
     // Prep the hitbox & Add it to the player
-    sf::RectangleShape hitboxShape = sf::RectangleShape(sf::Vector2f(this->uvRect.width * 0.1f, this->uvRect.height * 0.1f));
-    this->hitbox = new Hitbox(
-        // Shape
-        hitboxShape,
-        // Position
-        this->shape.getPosition(),
-        // Size
-        sf::Vector2f(this->uvRect.width * 0.1f, this->uvRect.height * 0.1f),
-        // Offset
-        sf::Vector2f( (hitboxShape.getSize().x / 2.f) -7.f, (hitboxShape.getSize().y / 2.f) - 4.f),
-        // Scale
-        sf::Vector2f(1.f, 1.f),
-        // Rotation
-        0.f,
-        // Debug
-        true
-    );
-    this->addComponent(hitbox);
+    // sf::RectangleShape hitboxShape = sf::RectangleShape(sf::Vector2f(this->uvRect.width * 0.1f, this->uvRect.height * 0.1f));
+    // this->hitbox = new Hitbox(
+    //     // Shape
+    //     hitboxShape,
+    //     // Position
+    //     this->shape.getPosition(),
+    //     // Size
+    //     sf::Vector2f(this->uvRect.width * 0.1f, this->uvRect.height * 0.1f),
+    //     // Offset
+    //     sf::Vector2f( (hitboxShape.getSize().x / 2.f) -7.f, (hitboxShape.getSize().y / 2.f) - 4.f),
+    //     // Scale
+    //     sf::Vector2f(1.f, 1.f),
+    //     // Rotation
+    //     0.f,
+    //     // Debug
+    //     true
+    // );
+    // this->addComponent(hitbox);
 }
 
 
@@ -78,7 +78,7 @@ const sf::Vector2f& Player::getPos() const { return this->shape.getPosition(); }
             this->move(deltaTime);
             this->mouseInteraction();
             // Update any variables inside the components
-            this->hitbox->position = this->shape.getPosition();
+            // this->hitbox->position = this->shape.getPosition();
 
             // Update the gameobject components
             this->GameObject::update(deltaTime);
