+++ engine/capulus_engine.hpp
@@ -1,0 +1,11 @@
+#ifndef CAPULUS_ENGINE_HPP
+#define CAPULUS_ENGINE_HPP
+
+namespace CapulusEngine {
+    // Leave empty because this is just a header file and the actual namespace code is in different files
+}
+
+#include "game.hpp"
+#include "gameobject.hpp"
+
+#endif
