+++ engine/components/ce_components.hpp
@@ -1,0 +1,14 @@
+#ifndef CE_COMPONENTS_HPP
+#define CE_COMPONENTS_HPP
+
+namespace CE_Components {
+    // Leave empty because this is just a header file and the actual namespace code is in different files
+}
+
+// Here is where we include all of the code that we want to be a part of the namespace
+#include "icomponent.hpp"
+#include "hitbox.hpp"
+#include "ce_sprite.hpp"
+#include "transform.hpp"
+
+#endif
