+++ src/game/IsleGame.hpp
@@ -1,0 +1,19 @@
+#ifndef ISLE_GAME_HPP
+#define ISLE_GAME_HPP
+
+
+
+namespace IsleGame {
+    // Empty since this is just a header
+}
+// Include the engine and engine components
+#include "../../engine/capulus_engine.hpp"
+using namespace CapulusEngine;
+using namespace CE_Components;
+
+// This is where we'll include all the custom objects we make
+#include "testObj.hpp"
+
+
+
+#endif
