+++ src/game/testObj.hpp
@@ -1,0 +1,24 @@
+#ifndef TESTOBJ_HPP
+#define TESTOBJ_HPP
+
+#include "IsleGame.hpp"
+
+namespace IsleGame {
+    class TestObj : public GameObject {
+        private:
+            // Variables
+        Sprite *sprite;
+        Hitbox *hitbox;
+
+        public:
+            // Constructor
+            TestObj();
+            // Destructor
+            virtual ~TestObj() override;
+
+            // Functions
+            void update(float frameTime);
+            void render(sf::RenderTarget *target);      
+    };
+}
+#endif
